# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AgroConsult API documentation",
        "description": "AgroConsult API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/response": {
            "get": {
                "tags": [
                    "Response"
                ],
                "summary": "response list",
                "description": "index",
                "operationId": "responseIndex",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Response"
                ],
                "summary": "Create new response",
                "description": "Create response",
                "operationId": "ec7153b0f00715e48d2a39172781a95b",
                "requestBody": {
                    "description": "Create new response",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "application_id",
                                    "resume_id",
                                    "amount",
                                    "text"
                                ],
                                "properties": {
                                    "application_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "resume_id": {
                                        "type": "number",
                                        "example": "3"
                                    },
                                    "amount": {
                                        "type": "text",
                                        "example": "250000"
                                    },
                                    "text": {
                                        "type": "text",
                                        "example": "backend dasturchi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, filled input. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/response/{response}": {
            "get": {
                "tags": [
                    "Response"
                ],
                "summary": "responseShow",
                "description": "Show by response",
                "operationId": "responseShow",
                "parameters": [
                    {
                        "name": "response",
                        "in": "path",
                        "description": "response ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Response"
                ],
                "summary": "Update response",
                "description": "Update response",
                "operationId": "b7a8c8d77cf1cf38a9c3708717f2260a",
                "parameters": [
                    {
                        "name": "response",
                        "in": "path",
                        "description": "response ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update Category",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "application_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "resume_id": {
                                        "type": "number",
                                        "example": "3"
                                    },
                                    "amount": {
                                        "type": "text",
                                        "example": "250000"
                                    },
                                    "text": {
                                        "type": "text",
                                        "example": "backend dasturchi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Response"
                ],
                "summary": "delete response",
                "description": "delete ",
                "operationId": "771c19b3c08ab2f39cf2f9fb03b697c3",
                "parameters": [
                    {
                        "name": "response",
                        "in": "path",
                        "description": "response ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with username and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}